pyproject.toml
[project]
name = "fw"
version = "0.1.0"
description = "Flow Wizard - A tool for workflow management"
requires-python = ">=3.8"
authors = [
    {name = "Jesper Vang", email = "jesper_vang@me.com"}
]
dependencies = [
    "typer",
    "rich",
    "questionary",
    "pyyaml",
    "gitignore-parser",
]

[project.scripts]
fw = "fw.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fw"]
